Refactor and rename stuff
