+++ web/src/street-explorer/stores.ts
@@ -2,4 +2,4 @@ import type { FeatureCollection } from "geojson";
 import { writable, type Writable } from "svelte/store";
 import { emptyGeojson } from "../common/utils";
 
-export const cycleGj: Writable<FeatureCollection> = writable(emptyGeojson());
+export const blockGj: Writable<FeatureCollection> = writable(emptyGeojson());
